Adding parentMs2 filed to ms2::Spectrum object.
Spectrum::precursorFile will not always be exactly the same as the actual parent ms2 file name.

This was causing problems with matching .spectrum files to .pdf files for the progressBar function in python/make_ms2.py

parentMs2 should be exactly the same as the ms2 file anme without the extesion.